+++ changelog.md
@@ -22,6 +22,8 @@ Codion Change Log
 - EntityEditModel.EntityEditor.ValueEditor renamed EditorValue.
 - DefaultEntityApplicationModel, entityModels now initialized via constructor parameter, related refactoring.
 - EntityQueryModel.LIMIT configuration value added.
+### is.codion.swing.framework.model
+- SwingEntityModel.Builder removed.
 ### is.codion.swing.framework.ui
 - EntityTablePanel.orderQueryBySortOrder() moved to model.
 - EntityPanel.Builder, major simplification and related refactoring.